tests added to doc provider
